Adding story
Night over so saving for day by day. Added story of one of the players and realized I wanted it to be in third person rather then first person like I wrote in my original story board. Shouldn't be to hard to change.

Also added an Enemy class so I can have Enemy's to fight. Simple attributes for enemys.